refactor: appConfig folder structure fix: eslint issues in appConfig components refactor: AppConfig.tsx add context provider
