+++ src/Pages/Applications/DevtronApps/Details/AppConfigurations/appConfig.scss
