+++ src/Pages/Applications/DevtronApps/Details/AppConfigurations/appConfig.service.ts
@@ -15,7 +15,7 @@
  */
 
 import { get, trash } from '@devtron-labs/devtron-fe-common-lib'
-import { Routes } from '../../../../config'
+import { Routes } from '../../../../../config'
 
 export const getGitProviderMin = () => {
     const URL = `${Routes.GIT_PROVIDER_MIN}`
