+++ src/components/app/typing.d.ts
@@ -15,3 +15,12 @@
  */
 
 declare module 'devtron_external_app'
+
+declare module '*.svg' {
+    import * as React from 'react'
+
+    export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>
+
+    const src: string
+    export default src
+}
