+++ src/components/deploymentConfig/types.tsx
@@ -18,7 +18,6 @@ import React from 'react'
 import { ResponseType, ServerError } from '@devtron-labs/devtron-fe-common-lib'
 import * as jsonpatch from 'fast-json-patch'
 import { AppEnvironment } from '../../services/service.types'
-import { CustomNavItemsType } from '../app/details/appConfig/appConfig.type'
 import { EnvironmentOverrideComponentProps } from '../EnvironmentOverride/EnvironmentOverrides.type'
 
 export interface DeploymentObject {
