+++ src/vite-env.d.ts
@@ -15,9 +15,10 @@
  */
 
 /// <reference types="react-scripts" />
-import 'jest-extended'
 /// <reference types="vite/client" />
 /// <reference types="vite-plugin-pwa/client" />
+
+import 'jest-extended'
 
 declare const __BASE_URL__: string
 declare const __ORCHESTRATOR_ROOT__: string
